app/api-gateway/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
astroid==3.2.4
attrs==24.2.0
black==24.8.0
certifi==2024.8.30
click==8.1.7
confluent-kafka==2.5.3
dill==0.3.8
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
isort==5.13.2
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pylint==3.2.7
python-telegram-bot==21.5
sniffio==1.3.1
tomlkit==0.13.2
yarl==1.9.11
